Add some defensive coding to overcome random 422 and 500 errors from IDM
These occurrences are completely random and usually retrying right after the failed call will result in a success.
This is of course highly undesirable in a Terraform pipeline so we
are willing to put up with uglifying the code a bit if it results
in a more robust execution experience for customers. However, this
is not sustainable on the long run and there really needs to be better
error handling in HSDP services (either internal retries, or better
error messages so we can do proper RCA) -- end rant

Signed-off-by: Andy Lo-A-Foe <andy.loafoe@gmail.com>