MDM: add retry to read operation as well
If MDM permissions are granted on-the-fly
then it could be that some read operations
happen on another thread which might not have
gotten a token refresh yet. We'll have to
investigate bit further to see if we should
share a single token across all the different
workers to alleviate this issue without retries
all over the place. On the other hand we
may not want to couple workers this tightly.

Signed-off-by: Andy Lo-A-Foe <andy.loafoe@gmail.com>